staging: sm7xxfb: remove driver
It hasn't been worked on in a very long time, and the original author
has moved on to a different product as this one is no longer being made.
So remove the driver.  If someone wants to resurect it, and clean it up
and get it merged to the "proper" part of the kernel, this commit can be
reverted.

Cc: Teddy Wang <teddy.wang@siliconmotion.com.cn>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>